fix incorrect reference to library function
While explaing the make_thread function with it's two paramters,
'pthread_create' is refered to instead of the wrapper function.
In later versions of the book this was erroneuosly attempted to fix
by commit `d0edd0e3` but it does not make sense with the context of the
previous paragraphs.